move part of route computation from RoutingListModel to Router, make computation asynchronous
