+++ libosmscout-client-qt/include/osmscout/Router.h
@@ -29,10 +29,78 @@
 
 #include <osmscout/private/ClientQtImportExport.h>
 
+Q_DECLARE_METATYPE(osmscout::Vehicle)
+
+/**
+ * \ingroup QtAPI
+ */
+class OSMSCOUT_CLIENT_QT_API RouteStep : public QObject
+{
+    Q_OBJECT
+    Q_PROPERTY(QString distance READ getDistance)
+    Q_PROPERTY(QString distanceDelta READ getDistanceDelta)
+    Q_PROPERTY(QString time READ getTime)
+    Q_PROPERTY(QString timeDelta READ getTimeDelta)
+    Q_PROPERTY(QString description READ getDescription)
+
+public:
+  QString distance;
+  QString distanceDelta;
+  QString time;
+  QString timeDelta;
+  QString description;
+
+public:
+  RouteStep();
+  RouteStep(const RouteStep& other);
+
+  RouteStep& operator=(const RouteStep& other);
+
+  QString getDistance() const
+  {
+      return distance;
+  }
+
+  QString getDistanceDelta() const
+  {
+      return distanceDelta;
+  }
+
+  QString getTime() const
+  {
+      return time;
+  }
+
+  QString getTimeDelta() const
+  {
+      return timeDelta;
+  }
+
+  QString getDescription() const
+  {
+      return description;
+  }
+};
+
+/**
+ * \ingroup QtAPI
+ */
+struct RouteSelection
+{
+  osmscout::RouteData        routeData;
+  osmscout::RouteDescription routeDescription;
+  QList<RouteStep>           routeSteps;
+};
+
+Q_DECLARE_METATYPE(RouteSelection)
+
+/**
+ * \ingroup QtAPI
+ */
 class OSMSCOUT_CLIENT_QT_API Router : public QObject{
   Q_OBJECT
 
-protected:
+private:
   QThread     *thread;
   SettingsRef settings;
   DBThreadRef dbThread;
@@ -40,6 +108,70 @@ protected:
 
 public slots:
   void Initialize();
+
+  /**
+   * Start Route computation. Router emits routeComputed or routeFailed later.
+   *
+   * User of this function should use Qt::QueuedConnection for invoking
+   * this slot, search may generate IO load and may tooks long time.
+   *
+   * Route computation can be long depending on the complexity of the route
+   * (the further away the endpoints, the more difficult the routing).
+   *
+   * @param start - starting position for route computation
+   * @param target - end position for route computation
+   * @param vehicle - used vehicle for route
+   * @param requestId - id used later in routeComputed/routeFailed signals
+   */
+  void onRouteRequest(LocationEntry* start,
+                      LocationEntry* target,
+                      osmscout::Vehicle vehicle,
+                      int requestId);
+signals:
+  void routeComputed(RouteSelection route,
+                     int requestId);
+
+  void routeFailed(QString reason,
+                   int requestId);
+
+private:
+  void GetCarSpeedTable(std::map<std::string,double>& map);
+
+  void DumpStartDescription(RouteSelection &route,
+                            const osmscout::RouteDescription::StartDescriptionRef& startDescription,
+                            const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
+
+  void DumpTargetDescription(RouteSelection &route, 
+                             const osmscout::RouteDescription::TargetDescriptionRef& targetDescription);
+
+  void DumpTurnDescription(RouteSelection &route,
+                           const osmscout::RouteDescription::TurnDescriptionRef& turnDescription,
+                           const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription,
+                           const osmscout::RouteDescription::DirectionDescriptionRef& directionDescription,
+                           const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
+
+  void DumpRoundaboutEnterDescription(RouteSelection &route,
+                                      const osmscout::RouteDescription::RoundaboutEnterDescriptionRef& roundaboutEnterDescription,
+                                      const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription);
+
+  void DumpRoundaboutLeaveDescription(RouteSelection &route,
+                                      const osmscout::RouteDescription::RoundaboutLeaveDescriptionRef& roundaboutLeaveDescription,
+                                      const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
+
+  void DumpMotorwayEnterDescription(RouteSelection &route,
+                                    const osmscout::RouteDescription::MotorwayEnterDescriptionRef& motorwayEnterDescription,
+                                    const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription);
+
+  void DumpMotorwayChangeDescription(RouteSelection &route,
+                                     const osmscout::RouteDescription::MotorwayChangeDescriptionRef& motorwayChangeDescription);
+  
+  void DumpMotorwayLeaveDescription(RouteSelection &route,
+                                    const osmscout::RouteDescription::MotorwayLeaveDescriptionRef& motorwayLeaveDescription,
+                                    const osmscout::RouteDescription::DirectionDescriptionRef& directionDescription,
+                                    const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
+  
+  void DumpNameChangedDescription(RouteSelection &route,
+                                  const osmscout::RouteDescription::NameChangedDescriptionRef& nameChangedDescription);
 
 public:
   Router(QThread *thread,
