+++ libosmscout-client-qt/include/osmscout/RoutingModel.h
@@ -37,126 +37,82 @@
 /**
  * \ingroup QtAPI
  */
-class OSMSCOUT_CLIENT_QT_API RouteStep : public QObject
-{
-    Q_OBJECT
-    Q_PROPERTY(QString distance READ getDistance)
-    Q_PROPERTY(QString distanceDelta READ getDistanceDelta)
-    Q_PROPERTY(QString time READ getTime)
-    Q_PROPERTY(QString timeDelta READ getTimeDelta)
-    Q_PROPERTY(QString description READ getDescription)
-
-public:
-  QString distance;
-  QString distanceDelta;
-  QString time;
-  QString timeDelta;
-  QString description;
-
-public:
-  RouteStep();
-  RouteStep(const RouteStep& other);
-
-  RouteStep& operator=(const RouteStep& other);
-
-  QString getDistance() const
-  {
-      return distance;
-  }
-
-  QString getDistanceDelta() const
-  {
-      return distanceDelta;
-  }
-
-  QString getTime() const
-  {
-      return time;
-  }
-
-  QString getTimeDelta() const
-  {
-      return timeDelta;
-  }
-
-  QString getDescription() const
-  {
-      return description;
-  }
-};
-
-/**
- * \ingroup QtAPI
- */
 class OSMSCOUT_CLIENT_QT_API RoutingListModel : public QAbstractListModel
 {
-    Q_OBJECT
-    Q_PROPERTY(int count READ rowCount)
+  Q_OBJECT
+  Q_PROPERTY(int count READ rowCount)
+  Q_PROPERTY(bool ready READ isReady NOTIFY computingChanged)
+
+signals:
+  void routeRequest(LocationEntry* start,
+                    LocationEntry* target,
+                    osmscout::Vehicle vehicle,
+                    int requestId);
+
+  void computingChanged();
+
+  void routeFailed(QString reason);
 
 public slots:
-    void setStartAndTarget(LocationEntry* start,
-                           LocationEntry* target);
-    void clear();
+  void setStartAndTarget(LocationEntry* start,
+                         LocationEntry* target,
+                         QString vehicleStr="car");
+
+  void clear();
+
+  void onRouteComputed(RouteSelection route,
+                       int requestId);
+
+  void onRouteFailed(QString reason,
+                     int requestId);
 
 private:
-    Router *router;
-    struct RouteSelection
-    {
-      osmscout::RouteData        routeData;
-      osmscout::RouteDescription routeDescription;
-      QList<RouteStep>           routeSteps;
-    };
-
-    RouteSelection route;
+  Router          *router;
+  RouteSelection  route;
+  int             requestId;
+  bool            computing;
 
 public:
-    enum Roles {
-        LabelRole = Qt::UserRole
-    };
-
-private:
-    void GetCarSpeedTable(std::map<std::string,double>& map);
-
-    void DumpStartDescription(const osmscout::RouteDescription::StartDescriptionRef& startDescription,
-                              const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
-    void DumpTargetDescription(const osmscout::RouteDescription::TargetDescriptionRef& targetDescription);
-    void DumpTurnDescription(const osmscout::RouteDescription::TurnDescriptionRef& turnDescription,
-                             const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription,
-                             const osmscout::RouteDescription::DirectionDescriptionRef& directionDescription,
-                             const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
-    void DumpRoundaboutEnterDescription(const osmscout::RouteDescription::RoundaboutEnterDescriptionRef& roundaboutEnterDescription,
-                                        const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription);
-    void DumpRoundaboutLeaveDescription(const osmscout::RouteDescription::RoundaboutLeaveDescriptionRef& roundaboutLeaveDescription,
-                                        const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
-    void DumpMotorwayEnterDescription(const osmscout::RouteDescription::MotorwayEnterDescriptionRef& motorwayEnterDescription,
-                                      const osmscout::RouteDescription::CrossingWaysDescriptionRef& crossingWaysDescription);
-    void DumpMotorwayChangeDescription(const osmscout::RouteDescription::MotorwayChangeDescriptionRef& motorwayChangeDescription);
-    void DumpMotorwayLeaveDescription(const osmscout::RouteDescription::MotorwayLeaveDescriptionRef& motorwayLeaveDescription,
-                                      const osmscout::RouteDescription::DirectionDescriptionRef& directionDescription,
-                                      const osmscout::RouteDescription::NameDescriptionRef& nameDescription);
-    void DumpNameChangedDescription(const osmscout::RouteDescription::NameChangedDescriptionRef& nameChangedDescription);
+  enum Roles {
+      LabelRole = Qt::UserRole
+  };
 
 public:
-    RoutingListModel(QObject* parent = 0);
-    virtual ~RoutingListModel();
+  RoutingListModel(QObject* parent = 0);
+  virtual ~RoutingListModel();
 
-    QVariant data(const QModelIndex &index, int role) const;
+  QVariant data(const QModelIndex &index, int role) const;
 
-    int rowCount(const QModelIndex &parent = QModelIndex()) const;
+  int rowCount(const QModelIndex &parent = QModelIndex()) const;
 
-    Qt::ItemFlags flags(const QModelIndex &index) const;
+  Qt::ItemFlags flags(const QModelIndex &index) const;
 
-    QHash<int, QByteArray> roleNames() const;
+  QHash<int, QByteArray> roleNames() const;
 
-    Q_INVOKABLE RouteStep* get(int row) const;
+  Q_INVOKABLE RouteStep* get(int row) const;
 
-    /**
-     * Create LocationEntry from geographic coordinate with optional label.
-     * It may be used from QML when selecting route start/end via point on map.
-     */
-    inline Q_INVOKABLE LocationEntry* locationEntryFromPosition(double lat, double lon, QString label=""){
-      return new LocationEntry(label,osmscout::GeoCoord(lat,lon));
-    }
+  inline bool isReady()
+  {
+    return !computing;
+  }
+
+  inline Q_INVOKABLE QStringList availableVehicles()
+  {
+    QStringList vehicles;
+    vehicles << "car";
+    vehicles << "bicycle";
+    vehicles << "foot";
+    return vehicles;
+  }
+
+  /**
+   * Create LocationEntry from geographic coordinate with optional label.
+   * It may be used from QML when selecting route start/end via point on map.
+   */
+  inline Q_INVOKABLE LocationEntry* locationEntryFromPosition(double lat, double lon, QString label="")
+  {
+    return new LocationEntry(label,osmscout::GeoCoord(lat,lon));
+  }
 };
 
 #endif
